added gulp. moved styles to styles folder
